% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{authors}
\alias{authors}
\title{authors}
\format{
\describe{
  \item{problem_id}{Unique identifier of the problem}
  \item{team_id}{Unique identifier for the team that produced the report}
  \item{team}{Name of the team that produced the report}
  \item{response_id}{Unique identifier for the a response}
  \item{user_id}{Unique identifier of a user who (co-)authored the response}
  \item{user}{The screen name of the (co-)author (at the time of response or comment creation)}
  \item{problem}{Short title of the problem}
}
}
\description{
Responses (i.e. reports or ""resources"") can have multiple authors.  This table captures the authors of responses.
}
