% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chat}
\alias{chat}
\title{chat}
\format{
\describe{
  \item{id}{Unique identifier of the chat message}
  \item{problem_id}{Unique identifier of the problem}
  \item{created_at}{Timestamp when the comment was first created}
  \item{user}{Name of the author of the chat message}
  \item{chat_id}{Another unique (within this platform instance) identifier of the chat message}
  \item{chat_text}{The comment content in plain text format}
  \item{user_id}{Unique identifier of the author of the chat message}
  \item{team_id}{Unique identifier for the team that produced the report}
  \item{team}{Name of the team the author of the chat message was in}
  \item{problem}{Short title of the problem}
}
}
\description{
Users can chat freely within each problem.  The following table captures the chat data.
}
