% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{comments}
\alias{comments}
\title{comments}
\format{
\describe{
  \item{problem_id}{Unique identifier of the problem}
  \item{team_id}{Unique identifier for the team that produced the report}
  \item{team}{Name of the team that produced the report}
  \item{response_id}{Unique identifier for the a response}
  \item{author_id}{Unique identifier of the original author of the response}
  \item{author}{Name of the orginal author of the response}
  \item{comment_id}{Unique identifier for the comment}
  \item{commenter_id}{Unique identifier of a user who created the comment}
  \item{commenter_name}{Name of the commenter}
  \item{time_created}{Timestamp when the comment was first created}
  \item{time_last_changed}{Timestamp when the comment was last updated}
  \item{comment_text}{A unique filename of the comment content in markdown text format}
  \item{comment_file}{A unique filename in the ""Comments"" subfolder that contains the HTML file of the comment}
  \item{problem}{Short title of the problem}
}
}
\description{
Users can comment on responses.  The following table captures these comments.
}
