% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{compile_data}
\alias{compile_data}
\title{Compile data from 'raw' SWARM, Qualtrics & Knack CSVs}
\usage{
compile_data(path = "HuntLab/data/")
}
\description{
\code{compile_data} Compiles data from 'raw' CSVs exported from the various
platforms used (SWARM, Qualtrics, Knack) for the Hunt Challenge 2020,
and saves tidied versions of them to the package data file, overwriting any
that were already saved.
}
\details{
Use this function to refresh the tidied versions of the data whenever the
raw data is updated.
}
\examples{
\dontrun{
# (re-)compile data
compile_data()
}
}
