% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-report-plots.R
\name{generate_feedback_plots}
\alias{generate_feedback_plots}
\title{Generate plots for feedback reports to organisations}
\usage{
generate_feedback_plots(instance_name)
}
\description{
\code{generate_feedback_plots} This function generates the graphs
that are to be used in the feedback reports to organisations 
following the completion of the 2020 Hunt Challenge.
}
\details{
The graphs will be generated as images in the 'exports' folder in
your R working directory. The images then need to be manually
inserted into the report template, which is a Google Doc. The 
template is called 'Organisation Feedback Template' and is stored
in the drive under:
  SWARM2_ALL
   > Hunt Challenge 2020
      > Reports to Organisations
}
\examples{
\dontrun{
# (re-)compile data
generate_feedback_plots('2020_HuntChallenge')
}
}
