% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HuntLab-package.R
\docType{package}
\name{HuntLab}
\alias{HuntLab}
\alias{HuntLab-package}
\title{HuntLab}
\description{
This package contains data and analysis functions for The Hunt
Laboratory for Intelligence Research at The University of Melbourne.
}
\details{
In particular, it contains data from the 2018 and 2020 challenges.
}
