% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huntr-package.R
\docType{package}
\name{huntr}
\alias{huntr}
\title{huntr}
\description{
This package contains data and analysis functions for The Hunt
Laboratory for Intelligence Research at The University of Melbourne.
}
\details{
In particular, it contains data from the 2018 and 2020 challenges, contained
in a an object called \code{repo}.

For more details see the Platform vignette:
\code{vignette("PlatformData", package = "huntr")}
}
