% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{login}
\alias{login}
\title{login}
\format{
\describe{
  \item{user_id}{Unique identifier of the user who logs in/out}
  \item{user}{Name of user who logs in/out}
  \item{event_type}{Can either be login or logout}
  \item{timestamp}{Timestamp when login/logout occurred}
}
}
\description{
We capture when users log in and out of the platform. A few caveats here: Since many users do not expressly logout but rather just close the browser window we can track far less logouts than logins (rougly a third). We have not attempted to estimate how long a user stays on platform. Also, we do not track which problem a user is looking at. This data can only be supplied on platform level. It includes all login/logouts to this environment, not scaled down to specific tests or exercises.
}
