% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{problems}
\alias{problems}
\title{problems}
\format{
\describe{
  \item{problem_id}{Unique identifier of the problem}
  \item{problem}{Short title of the problem}
  \item{description}{}
  \item{start_date}{}
  \item{due_date}{}
  \item{problem_text}{A unique filename of the problem content in markdown text format}
  \item{problem_file}{A unique filename in the ""Problems"" subfolder that contains the HTML file of the problem}
  \item{problem_template}{}
  \item{exercise}{}
}
}
\description{
The problems table uniquely identifies and describes the problem that SWARM users were working on.
}
