% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{probparts}
\alias{probparts}
\title{probparts}
\format{
\describe{
  \item{team}{Name of team the user is in}
  \item{user}{Name of user}
  \item{report_count}{Number of reports user contributed to problem}
  \item{resource_count}{Number of resources user contributed to problem}
  \item{comment_count}{Number of comments user contributed in a problem}
  \item{chat_count}{Number of chat messages user contributed in a problem}
  \item{comment_vote_count}{Number of comment upvotes user contributed in a problem}
  \item{resource_vote_count}{Number of resource upvotes user contributed in a problem}
  \item{simple_rating}{Number of times a user contributed a simple (single number) rating}
  \item{partial_rating}{Number of times a user contributed a partial (some of the rubrics) rating}
  \item{complete_rating}{Number of times a user contributed a complete (all rubrics) rating}
  \item{problem}{Title of the problem}
  \item{engagement}{Calculated number of overall engagement of one user to a problem}
  \item{report_count_scaled}{Number of reports user contributed to problem, scaled (within this platform instance) to range between 0 and 1}
  \item{resource_count_scaled}{Number of resources user contributed to problem, scaled (within this platform instance) to range between 0 and 1}
  \item{comment_count_scaled}{Number of comments user contributed in a problem, scaled (within this platform instance) to range between 0 and 1}
  \item{chat_count_scaled}{Number of chat messages user contributed in a problem, scaled (within this platform instance) to range between 0 and 1}
  \item{comment_vote_count_scaled}{Number of comment upvotes user contributed in a problem, scaled (within this platform instance) to range between 0 and 1}
  \item{resource_vote_count_scaled}{Number of resource upvotes user contributed in a problem, scaled (within this platform instance) to range between 0 and 1}
  \item{simple_rating_scaled}{Number of times a user contributed a simple (single number) rating, scaled (within this platform instance) to range between 0 and 1}
  \item{partial_rating_scaled}{Number of times a user contributed a partial (some of the rubrics) rating, scaled (within this platform instance) to range between 0 and 1}
  \item{complete_rating_scaled}{Number of times a user contributed a complete (all rubrics) rating, scaled (within this platform instance) to range between 0 and 1}
  \item{engagement_scaled}{Calculated number of overall engagement of one user to a problem, using scaled engagement scores as inputs}
  \item{vote_count}{comment_vote_count + resource_vote_count}
  \item{quick_rating}{simple_rating + partial_rating}
  \item{cluster}{Integer ID for the nearest ‘user role’ cluster}
  \item{clusterLabel}{Descriptive label for the nearest ‘user role’ cluster}
}
}
\description{
One row per participant-problem pairing (instance of a participant being a member of a team attempting a particular problem), with related attributes.
}
