% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ratings}
\alias{ratings}
\title{ratings}
\format{
\describe{
  \item{problem_id}{Unique identifier of the problem}
  \item{team_id}{Unique identifier for the team that produced the report}
  \item{content_type}{""report"", ""resource"", or ""comment"". Other types may be added in future.}
  \item{content_id}{Unique identifier for the content (content can be response or comment)}
  \item{author_id}{Unique identifier of the original author of the response}
  \item{author}{Name of the user who originally authored the response}
  \item{rater_id}{Unique identifier of a user who rated the response}
  \item{rater}{Name of the user who rated the response}
  \item{last_edited}{Timestamp when the rating was last updated}
  \item{rating_type}{""thumbs"", ""simple"", ""detailed""}
  \item{rating_total}{1 or 0 for thumbs, number (0-100) for simple, aggregated number for detailed rating}
  \item{rating_sub_01_title}{The title of the ratings subfield}
  \item{rating_sub_01}{Number for detailed rating field 1}
  \item{rating_sub_02_title}{The title of the ratings subfield}
  \item{rating_sub_02}{Number for detailed rating field 2}
  \item{rating_sub_03_title}{The title of the ratings subfield}
  \item{rating_sub_03}{Number for detailed rating field 3}
  \item{rating_sub_04_title}{The title of the ratings subfield}
  \item{rating_sub_04}{Number for detailed rating field 4}
  \item{rating_sub_05_title}{The title of the ratings subfield}
  \item{rating_sub_05}{Number for detailed rating field 5}
  \item{rating_sub_06_title}{The title of the ratings subfield}
  \item{rating_sub_06}{Number for detailed rating field 6}
  \item{rating_sub_07_title}{The title of the ratings subfield}
  \item{rating_sub_07}{Number for detailed rating field 7}
  \item{rating_sub_08_title}{The title of the ratings subfield}
  \item{rating_sub_08}{Number for detailed rating field 8}
  \item{rating_sub_09_title}{The title of the ratings subfield}
  \item{rating_sub_09}{Number for detailed rating field 9}
  \item{problem}{Short title of the problem}
}
}
\description{
Users can rate the responses and comments of other users.  The following table captures these ratings.  Note that users can rate as many times as they like (each of which is captured in the table), but only the last ratings are used in calculating the aggregate rating.  In other words, the ""rating_total"" with the latest timestamp in the *Ratings* table should match the ""team_rating"" in the *TopReport* table.
}
