% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{relations}
\alias{relations}
\title{relations}
\format{
\describe{
  \item{addressant}{User addressed by the other user (e.g. mentioned in chat)}
  \item{problem_id}{Unique identifier of problem}
  \item{team}{Name of team both users are in}
  \item{user}{User who addresses another user ( e.g. writes a chat message that contains a name tag)}
  \item{problem}{Name of problem}
  \item{interaction}{Describes the interaction: 'comment', 'contributor' or 'chat'}
  \item{count}{Number of times this interaction happened}
}
}
\description{
A table collecting interaction between users. As an interaction counts if a user commented on another users' response, if a user is named as a contributor for a report or if a user mentions another user in the chat. The interactions are stored summed up per problem and per action.
}
