% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{responses}
\alias{responses}
\title{responses}
\format{
\describe{
  \item{problem_id}{Unique identifier of the problem}
  \item{team_id}{Unique identifier for the team that produced the response}
  \item{team}{The name of the team that produced the response}
  \item{response_id}{Unique identifier for the a response}
  \item{time_created}{Timestamp when the response was first created}
  \item{time_last_changed}{Timestamp when the response was last updated}
  \item{response_title}{The beginning of the response in plain text format}
  \item{response_type}{Type of response, currently 'report' or 'resource'. More types may be added in future}
  \item{response_text}{A unique filename of the response content in markdown text format}
  \item{response_file}{A unique filename in the 'Responses' subfolder that contains the HTML file of the response}
  \item{problem}{Short title of the problem}
}
}
\description{
Responses are text contributed by users in response to the problem.  Currently, responses can be ""reports"" or ""resources"", but more types are possible in future.
}
