% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{teamparts}
\alias{teamparts}
\title{teamparts}
\format{
\describe{
  \item{user}{Username.}
  \item{team}{A team of which the user was a member. (Users may be members of multiple teams over the course of an exercise. For example, they may be in an ordinary public team early on, and then get selected for a superteam later. In such a case, there would be two rows in this table, one for each pairing of a user with a team.)}
}
}
\description{
One row per team-participant pairing. Useful for pairing participants with teams. Participants can be members of multiple teams over the course of an experiment. For example, a user may be a member of a generic public team in the beginning, and then be selected for a super team later on.
}
