% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{teams}
\alias{teams}
\title{teams}
\format{
\describe{
  \item{team}{Team name.}
  \item{AOMT}{Median Actively Open-Minded Thinking score for team members. Missing values are excluded from the calculation.}
  \item{divAOMT}{Average pairwise difference in AOMT scores across all team members for whom the AOMT score is available.}
  \item{medianEdu}{Median education level for the team, where education levels are encoded as follows. 1 = ‘High School’; 2 = ‘Trade or Technical Qualification’; 3 = ‘Bachelors’; 4 = ‘Graduate Certificate, Diploma or equivalent’; 5 = ‘Masters’; 6 = ‘Phd’. Missing values excluded from calculation.}
  \item{type}{Team type. PT = Public Team; ST = Superteam; OT = Organisational Team;}
}
}
\description{
One row per team in the competition, with related attributes.
}
