% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{top_reports}
\alias{top_reports}
\title{top_reports}
\format{
\describe{
  \item{problem_id}{Unique identifier of the problem}
  \item{problem}{Short title of the problem}
  \item{team_id}{Unique identifier for the team that produced the report}
  \item{team}{Screen name for the team_id}
  \item{report_id}{Unique identifier for the report (response of type ""report"")}
  \item{team_rating}{The aggregate rating of the report from the (internal) team at time of submission}
  \item{rating_count}{Number of received ratings}
  \item{origin_author_id}{Unique identifier of the user that first created the response that ultimately resulted in this top report}
  \item{origin_author_name}{Name of the author that first created the response}
}
}
\description{
Top Reports are the reports that have the highest quality (readiness) rating at the time of the data dump. Of particular interest are ""final reports"" at the end of a test, but it is possible that an interim data dump is made, in which case the top report is the highest rated report at the time of the data set. Internally, it is also possible that such reports are identifiable through other data in the SWARM platform (such as a report submission event log), but only the basic information is captured in the table below.
}
