% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{parts}
\alias{parts}
\title{parts}
\format{
\describe{
  \item{user}{Username.}
  \item{isOrg}{Logical. Equals TRUE if user participated as part of an organisational team.}
  \item{gaveConsent2}{Gave consent to participate in the ‘Problem Solving In Online Groups’ experiment.}
  \item{gaveConsent3}{Gave consent to participate in the ‘Identifying and Rating Quality of Reasoning’ experiment.}
  \item{gaveConsent4}{Gave consent to proceed with the Entry Survey in the Psychology Capstone exercise.}
  \item{agreedToTerms}{}
  \item{agegroup}{Age group.}
  \item{gender}{Gender.}
  \item{occupation}{Occupation status.}
  \item{education}{Highest education level.}
  \item{major1, major2}{Subjects majored in.}
  \item{minor1, minor2}{Subjects minored in.}
  \item{studyarea}{Study area.}
  \item{studyareaOtherInput}{Free text input for ‘Other’ study area, if selected.}
  \item{englishProficiency}{Proficiency in English. Response options are: ‘No proficiency’, ‘Elementary proficiency’, ‘Limited working proficiency’, ‘Professional working proficiency’, ‘Full professional proficiency’, ‘Native or bilingual proficiency’}
  \item{loteProficiency1-6}{Specified proficiency levels in up to 6 languages other than English.}
  \item{loteProficiencyText1-6}{The languages for which the proficiency is specified in loteProficiency.}
  \item{enjoyLogicProbs}{Response to the prompt: ‘How enjoyable do you find logic problems?’ Response options are: ‘Note at all enjoyable’, ‘A little bit enjoyable’, ‘Slightly enjoyable’, ‘Somewhat enjoyable’, ‘Quite enjoyable’, ‘Very enjoyable’, ‘Extremely enjoyable’}
  \item{enjoyNumProbs}{Response to the prompt: ‘How enjoyable do you find number problems?’ Response options are: ‘Note at all enjoyable’, ‘A little bit enjoyable’, ‘Slightly enjoyable’, ‘Somewhat enjoyable’, ‘Quite enjoyable’, ‘Very enjoyable’, ‘Extremely enjoyable’}
  \item{exp1}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Mathematics (generally)’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp2}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Quantitative  Modeling, Simulation’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp3}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Statistics’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp4}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Probability’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp5}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Bayes Nets’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp6}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Programming’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp7}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Experimental Design’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp8}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Risk Analysis’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp9}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Forecasting’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp10}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Decision Thoery (e.g. multi-attribute utility theory)’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp11}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Game Theory’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp12}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Structured Analytic Techniques’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp13}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Argument Mapping’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp14}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Informal Logic’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp15}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Systems Thinking or Cognitive Mapping’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp16}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Image Analysis’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp17}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Link Analysis’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp18}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Graphic Design’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{exp19}{Numerical encoding of the participant’s response to the prompt: ‘Please rate your level of technical expertise in the following areas: Technical Writing’. 1 = ‘Not familiar with this domain’, 2 = ‘Studied in school, but don’t use it’, 3 = ‘Use this knowledge occasionally’, 4 = ‘Use this knowledge regularly’, 5 = ‘I am a recognized expert’, 6 = ‘I am an international authority’.}
  \item{pc…}{Probability coherence questions. The prompt is: ‘What is the probability that each of the following statements are true? If you are absolutely certain that the statement is true, you should enter 100. Likewise, if you are absolutely certain that the statement is false, you should enter 0. If you are uncertain, you should enter the probability that corresponds with what you think are the chances that the statement is true. When answering these questions, we ask that you do NOT consult with any outside sources.’ Respondents then select a probability as a percentage.}
  \item{pc5A}{'HA is the postal abbreviation for Hawaii.'}
  \item{pc5B}{'HI is the postal abbreviation for Hawaii.'}
  \item{pc5Acomp}{'HA is NOT the postal abbreviation for Hawaii.'}
  \item{pc5AUB}{'Either HA or HI is the postal abbreviation for Hawaii.'}
  \item{pc14A}{'Michaelangelo painted the Sistine Chapel.'}
  \item{pc14B}{'Raphael painted the Sistine Chapel.'}
  \item{pc14Acomp}{'Michaelangelo did NOT paint the Sistine Chapel.'}
  \item{pc14AUB}{'Either Michaelangelo or Raphael painted the Sistine Chapel.'}
  \item{pc35A}{'The Pacific Ocean is the largest of Earth’s oceans.'}
  \item{pc35B}{'The Indian Ocean is the largest of Earth’s oceans.'}
  \item{pc35Acomp}{'The Pacific Ocean is NOT the largest of the Earth’s oceans.'}
  \item{pc35AUB}{'Either the Pacific Ocean or the Indian Ocean is the largest of Earth’s oceans.'}
  \item{pc57A}{'The term 'bounded rationality' was coined by Herbert Simon.'}
  \item{pc57B}{'The term 'bounded rationality' was coined by Alan Turing.'}
  \item{pc57Acomp}{'The term 'bounded rationality' was NOT coined by Herbert Simon.'}
  \item{pc57AUB}{'The term 'bounded rationality' was coined either by Herbert Simon or Alan Turing.'}
  \item{pc60A}{'The oldest brother of famous philosopher William James was famous politician Walter James.'}
  \item{pc60B}{'The oldest brother of famous philosopher William James was famous writer Henry James.'}
  \item{pc60Acomp}{'The oldest brother of famous philosopher William James was NOT the famous politician Walter James.'}
  \item{pc60AUB}{'The oldest brother of famous philosopher William James was either famous politician Walter James or famous writer Henry James.'}
  \item{pr1}{Probabilistic reasoning item. Prompt: 'A ball was drawn from a bag ontaning 10 red, 30 white, 20 blue, and 15 yellow balls. What is the probability that it is neither red nor blue?' Response options: '30/75', '10/75', '45.75'.}
  \item{pr2}{Probabilistic reasoning item. Prompt: 'The proportion of left-handed people is 4 out of 100. In a school there are 300 students. How many left-handed students are there?' Response options: 20, 12, 25.}
  \item{pr3}{Probabilistic reasoning item. Prompt: 'Smokers are about 35\% of the population. At the airport in the waiting room there are 200 passengers. How many smokers are there?' Response options: 70, 35, 20.}
  \item{pr4}{Probabilistic reasoning item. Prompt: 'A fair coin is tossed nine times. Which of the following sequences of outcomes is a more likely result of nine flips of the fair coin? (H: Heads, T: Tails)'. Response options: 'THHTHTTHH', 'HTHTHTHTH', 'Both sequences are equally likely'}
  \item{pr5}{Probabilistic reasoning item. Prompt: 'Two containers, labelled A and B, are filled with red and yellow tokens in the following quantities. Container A contains 100 tokens, 65 red and 35 yellow. Container B contains 10 tokens, 6 red and 4 yellow. Each container is shaken vigorously. After choosing one of the containers, you must draw a token (without peeking, of course). Which container gives you a better chance of drawing a yellow token?' Response options: 'Container A (with 65 red and 35 yellow)', 'Container B (with 6 red and 4 yellow)', 'Equal chances from each container'.}
  \item{pr6}{Probabilistic reasoning item. Prompt: 'A marble bag contains 15 blue and 15 green marbles. After you drew 5 marbles (the marble draw was always put back into the bag), a sequence of five green marbles was obtained. What is the most likely outcome if a marble is drawn a sixth time?' Response options: 'A green marble', 'A blue marble', 'Blue and green are equally likely'.}
  \item{pr7}{Probabilistic reasoning item. Prompt: 'A bingo game is played with 25 numbers (from 1 to 25). At the first draw, which of the following results is the most likely?' Response options: 'It is more likely to be an even number', 'It is more likely to be an odd number', 'It is just as likely to be an even or an odd number'.}
  \item{pr8}{Probabilistic reasoning item. Prompt: 'Two decks, labelled A and B, are composed of cards with a star (star cards) and cards without any figure (white cards) on the reverse side. Deck A contains 100 cards, 80 white and 20 with a star. Deck B contains 10 cards, 8 white and 2 with a star. After choosing one of the decks, you must draw a card (without peeking, of course). Which deck gives you a better chance of drawing a star card?' Response options: 'Deck A (with 80 white and 20 star cards)', 'Deck B (with 8 white and 2 star cards', 'Equal chances from each deck'.}
  \item{pr9}{Probabilistic reasoning item. Prompt: 'A marble bag contains 10 blue and 20 green marbles. After you drew 5 marbles (the marble draw was always put back into the bag), a sequence of five green marbles was obtained. What is the most likely outcome if a marble is drawn a sixth time?' Response options: 'A green marble', 'A blue marble', 'Blue and green are equally likely'.}
  \item{pr10}{Probabilistic reasoning item. Prompt: '60\% of the population in a city are men and 40\% are women. 50\% of the men and 30\% of the women smoke. We select a person from the city at random. What is the probability that this person is a smoker?' Response options: 42\%, 50\%, 85\%.}
  \item{pr11}{Probabilistic reasoning item. Prompt: 'According to a recent survey, 90\% of the population in a city usually lie and 30\% of those usually lie about important matters. If we pick a person at random from this city, what is the probability that the person usually lies about important matters?' Response options: 60\%, 30\%, 27\%.}
  \item{pr12}{Probabilistic reasoning item. Prompt: 'In a choir there are 100 children: 30 boys and 70 girls. Half of the boys and 1 in 10 girls learn to play the piano. We select a child from the choir at random. What is the probability that he/she plays the piano?' Response options: '22 out of 100', '30 out of 100', '50 out of 100'.}
  \item{pr13}{Probabilistic reasoning item. Prompt: 'A village has 1000 inhabitants. 600 people own a pet, and amongst pet owners 1 in 3 owns more than one pet. If we select one person from this village at random, what is the probability that they own more than one pet' Response options: '333 out of 1000', '500 out of 1000', '200 out of 1000'.}
  \item{pr14}{Probabilistic reasoning item. Prompt: 'In a medical center a group of people were interviewed with the following results: (a table: ((,55 years old or younger,Over 55 years old, Total),(Previous heart attack, 29, 75, 104),(No previous heart attack, 401, 275, 676),(Total, 430, 350, 780))). Suppose we select a person from this group at random. Based on the table: What is the probability that the person has had a heart attack?' Response options: '104 out of 780', '104 out of 676', '390 out fo 780'.}
  \item{pr15}{Probabilistic reasoning item. Prompt: 'What is the probability that the person has had a heart attack and, at the same time is older than 55?' Response options: '104 out of 350', '75 out of 350', '75 out of 780'.}
  \item{pr16}{Probabilistic reasoning item. Prompt: 'When the person had a heart attack, what is the probability that they are over 55?' Response options: '75 out of 780', '75 out of 104', '104 out of 350'.}
  \item{mat1-mat11}{Responses to ICAR matrix reasoning questions.}
  \item{crt1}{Cognitive reflection task item. Prompt: 'A bat and a ball cost $1.10 in total. The bat costs a dollar more than the ball. How much does the ball cost? (Please enter a number in dollars, excluding the dollar sign.' }
  \item{crt2}{Cognitive reflection task item. Prompt: 'If it takes 5 machines 5 minutes to make 5 widgets, how long would it take for 100 machines to make 100 widgets? (Please enter a number in minutes)'}
  \item{crt3}{Cognitive reflection task item. Prompt: 'In a lake, there is a patch of lily pads. Every day, the patch doubles in size. If it takes 48 days for the patch to cover the entire lake, how long would it take for the patch to cover half the lake? (Please enter a number in days)'}
  \item{crt4}{Cognitive reflection task item. Prompt: 'If it takes 2 nurses 2 minutes to measure the blood pressure of 2 patients, how long would it take 200 nurses to measure the blood pressure of 200 patients? (Please enter a number in minutes)'}
  \item{crt5}{Cognitive reflection task item. Prompt: 'Soup and salad costs $5.50 in total. The soup costs a dollar more than the salad. How much does the salad cost? (Please enter a number in dollars, excluding the dollar sign)'}
  \item{crt6}{Cognitive reflection task item. Prompt: 'Sally is making sun tea. Every hour, the concentration of the tea doubles. If it takes 6 hours for the tea to be ready, how long would it take for the tea to reach half of the final concentration? (Please enter a number in hours)'}
  \item{crtSeenBefore}{Cognitive reflection task supplementary-item. 'Have you seen any of these 6 questions before?' Response options: Yes, No.}
  \item{crtSeenBeforeText}{Cognitive reflection task supplementary-item. 'Which ones?'}
  \item{bfi1}{Big Five Inventory item. Prompt: 'Below are a number of characteristics that may or may not apply to you. For example, do you agree that you are someone who likes to spend time with others? Please rate the extent to which you agree or disagree with each statement. I am someone who is reserved.' 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{bfi2}{Big Five Inventory item. Prompt: 'I am someone who is generally trusting.' 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{bfi3}{Big Five Inventory item. Prompt: 'I am someone who tends to be lazy.' 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{bfi4}{Big Five Inventory item. Prompt: 'I am someone who is relaxed, handles stress well.' 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{bfi5}{Big Five Inventory item. Prompt: 'I am someone who has few artistic interests.' 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{bfi6}{Big Five Inventory item. Prompt: 'I am someone who is outgoing, sociable.' 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{bfi7}{Big Five Inventory item. Prompt: 'I am someone who tends to find faults with others.' 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{bfi8}{Big Five Inventory item. Prompt: 'I am someone who does a thorough job.' 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{bfi9}{Big Five Inventory item. Prompt: 'I am someone who gets nervous easily.' 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{bfi10}{Big Five Inventory item. Prompt: 'I am someone who has an active imagination.' 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{tp1}{Teamwork Perceptions item. Prompt: 'I have found working as part of a team in my classes to be a valuable experience.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp2}{Teamwork Perceptions item. Prompt: 'In most of the teams I have been on, the other team members have generally contributed as much as I have.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp3}{Teamwork Perceptions item. Prompt: 'In most of the teams I have been on, I felt the other team members respected me.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp4}{Teamwork Perceptions item. Prompt: 'In most of the teams I have been on, the team has worked well together.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp5}{Teamwork Perceptions item. Prompt: 'I have found teamwork to be a productive use of course time.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp6}{Teamwork Perceptions item. Prompt: 'I have found that teams help me learn course material more than if I just studied alone.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp7}{Teamwork Perceptions item. Prompt: 'I have learned more in courses where I have been a member of a team.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp8}{Teamwork Perceptions item. Prompt: 'I have found being part of a team improved my course grades.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp9}{Teamwork Perceptions item. Prompt: 'I have found that working with a team helps me develop skills in working with others.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp10}{Teamwork Perceptions item. Prompt: 'I have found that working with a team has helped me develop cooperative leadership skills.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp11}{Teamwork Perceptions item. Prompt: 'I have found that working with a team has helped me develop more respect for the opinion of others.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp12}{Teamwork Perceptions item. Prompt: 'I have found that working with a team has enhanced my sense of who I am.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp13}{Teamwork Perceptions item. Prompt: 'I have found that teams make good decisions.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp14}{Teamwork Perceptions item. Prompt: 'I have found that being on a team has helped me become better at problem solving.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp15}{Teamwork Perceptions item. Prompt: 'Being part of a team discussion has improved my ability to think through a problem.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{tp16}{Teamwork Perceptions item. Prompt: 'I feel that team-based learning has improved my reasoning skills.' 1 = ‘Strongly disagree’, 2 = ‘Somewhat disagree’, 3 = ‘Neither agree nor disagree’, 4 = ‘Somewhat agree’, 5 = ‘Strongly agree’.}
  \item{int1}{Logical. Indicates whether the participant selected ‘What the problems will be like’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int2}{Logical. Indicates whether the participant selected ‘Platform functionality’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int3}{Logical. Indicates whether the participant selected ‘How reports are created’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int4}{Logical. Indicates whether the participant selected ‘Team collaboration experience’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int5}{Logical. Indicates whether the participant selected ‘The structured training available’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int6}{Logical. Indicates whether the participant selected ‘The tools in the Lens Kit’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int7}{Logical. Indicates whether the participant selected ‘The Contending Analyses methodology’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int8}{Logical. Indicates whether the participant selected ‘The evaluation methods’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int9}{Logical. Indicates whether the participant selected ‘How my team performs’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int10}{Logical. Indicates whether the participant selected ‘Whether the public do as well as the professionals’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int11}{Logical. Indicates whether the participant selected ‘Other’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{interestsOtherInput}{If the team selected ‘Other’ among their responses to the prompt ‘Thinking about the Hunt Challenge, what are you most interested in?’, this field contains their free text response.}
  \item{enExpct1}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Interesting problems’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct2}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Reasonable time commitment’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct3}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Difficult problems’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct4}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: I will learn new skills and tools’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct5}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Problems are achievable, for a team’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct6}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: The platform will be a productive work space’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct7}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Training in analytical techniques’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct8}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Positive team working experience’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct9}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: An effective collaboration, compared to my normal methods’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct10}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: What I learn from the challenge can be applied in my workplace’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{pri1}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Seeing how my team performs against the others’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri2}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Testing my own skills and abilities’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri3}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - I think it will be fun’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri4}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Developing my overall analytic skills’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri5}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Learning how the SWARM Platform works’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri6}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Learning about Contending Analyses’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri7}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Trying a new style of team collaboration’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri8}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Contributing to research in intelligence analysis’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri9}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Getting selected for a Super Team’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri10}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Exploring the Lens Kit’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri11}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Participating in crowd sourced intelligence analysis’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{priOther}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Other’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{priOtherInput}{Specified ‘Other’ priority referred to by priOther (if applicable)}
  \item{aomt1}{Numerical encoding of the participants' response to the prompt: ‘Allowing oneself to be convinced by an opposing argument is a sign of good character.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt2}{Numerical encoding of the participants' response to the prompt: ‘People should take into consideration evidence that goes against their beliefs.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt3}{Numerical encoding of the participants' response to the prompt: ‘People should revise their beliefs in light of new information or evidence.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt4}{Numerical encoding of the participants' response to the prompt: ‘Changing your mind is a sign of weakness.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt5}{Numerical encoding of the participants' response to the prompt: ‘Intuition is the best guide in making decisions.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt6}{Numerical encoding of the participants' response to the prompt: ‘It is important to persevere in your beliefs even when evidence is brought to bear against them.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt7}{Numerical encoding of the participants' response to the prompt: ‘One should disregard evidence that conflicts with one's established beliefs.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt8}{Numerical encoding of the participants' response to the prompt: ‘People should search actively for reasons why their beliefs might be wrong.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt9}{Numerical encoding of the participants' response to the prompt: ‘When we are faced with a new question, the first answer that occurs to us is usually best.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt10}{Numerical encoding of the participants' response to the prompt: ‘When faced with a new question, we should consider more than one possible answer before reaching a conclusion.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt11}{Numerical encoding of the participants' response to the prompt: ‘When faced with a new question, we should look for reasons why our first answer might be wrong, before deciding on an answer.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{enCap1}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Analytic report writing’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap2}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Using structured analytic techniques’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap3}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Using OSINT tools and resources’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap4}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Applying strategic thinking frameworks’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap5}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Identifying and analysing assumptions’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap6}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Evaluating quality of analytic reasoning’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap7}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Using decision making frameworks’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{yearsWorkExperience}{Response to the prompt: ‘How many years work experience do you have?’}
  \item{yearsAnalyticalExperience}{Response to the prompt: ‘How many years experience do you have analysing complex problems or data?’}
  \item{ae1}{Logical. Indicates whether the participant selected ‘No direct experience’ among their responses to the prompt: ‘Do you have experience analysing complex problems or data? Please select all applicable.’.}
  \item{ae2}{Logical. Indicates whether the participant selected ‘Yes, in an intelligence or related field’ among their responses to the prompt: ‘Do you have experience analysing complex problems or data? Please select all applicable.’.}
  \item{ae3}{Logical. Indicates whether the participant selected ‘Yes, in a scientific field’ among their responses to the prompt: ‘Do you have experience analysing complex problems or data? Please select all applicable.’.}
  \item{ae4}{Logical. Indicates whether the participant selected ‘Yes, in another field’ among their responses to the prompt: ‘Do you have experience analysing complex problems or data? Please select all applicable.’.}
  \item{ae5}{Logical. Indicates whether the participant selected ‘Prefer not to say’ among their responses to the prompt: ‘Do you have experience analysing complex problems or data? Please select all applicable.’.}
  \item{hasMultidisciplinaryExperience}{Logical. Indicates whether the participant responded ‘Yes’ to the prompt: ‘Have you had experience working in a multidisciplinary team? (i.e. composed of members with varied but complementary experience, qualifications, and skills that contribute to the achievement of a specific objective)’}
  \item{multidisciplinaryExperienceInput}{Response to the prompt: ‘Please describe your experience working in a multidisciplinary team:’}
  \item{starRating}{Integer. Number of stars selected in response to the prompt: ‘Please rate your overall Challenge experience out of 5 stars:’}
  \item{timeWellSpent}{Numerical encoding of the participants' response to the prompt: ‘Did you feel that your participation was time well spent?’. 1 = ‘No’, 2 = ‘Unsure’, 3 = ‘Yes’}
  \item{bestThing}{Response to the prompt: ‘What was the one BEST thing about your Challenge experience?’}
  \item{worstThing}{Response to the prompt: ‘What was the one WORST thing about your Challenge experience?’}
  \item{hoursPerWeek}{Response to the prompt: ‘How many hours did your organisation allow for your participation during work hours per week? Please enter the number, 1, 2, 3, etc...’}
  \item{enoughtTime}{Numerical encoding of the participants' response to the prompt: ‘Did your organisation allow you enough time off regular duties to participate fully?’. 1 = ‘No’, 2 = ‘Yes’}
  \item{proportionOwnTime}{Response to the prompt: ‘What proportion of your participation was done in your own time? (outside of work hours)? Please enter the number, 10, 25, 50, etc...’}
  \item{rate1}{Numerical encoding of the participants' response to the prompt: ‘How do you rate these aspects of the Challenge? - The onboarding process’. 1 = ‘Poor’, 2 = ‘Average’, 3 = ‘Good’}
  \item{rate2}{Numerical encoding of the participants' response to the prompt: ‘How do you rate these aspects of the Challenge? - Our communication with you’. 1 = ‘Poor’, 2 = ‘Average’, 3 = ‘Good’}
  \item{rate3}{Numerical encoding of the participants' response to the prompt: ‘How do you rate these aspects of the Challenge? - The training’. 1 = ‘Poor’, 2 = ‘Average’, 3 = ‘Good’}
  \item{rate4}{Numerical encoding of the participants' response to the prompt: ‘How do you rate these aspects of the Challenge? - The feedback’. 1 = ‘Poor’, 2 = ‘Average’, 3 = ‘Good’}
  \item{rate5}{Numerical encoding of the participants' response to the prompt: ‘How do you rate these aspects of the Challenge? - The Help Center’. 1 = ‘Poor’, 2 = ‘Average’, 3 = ‘Good’}
  \item{exExpct1}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Interesting problems’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct2}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Reasonable time commitment’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct3}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Difficult problems’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct4}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: I will learn new skills and tools’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct5}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Problems are achievable, for a team’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct6}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: The platform will be a productive work space’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct7}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Training in analytical techniques’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct8}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Positive team working experience’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct9}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: An effective collaboration, compared to my normal methods’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct10}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: What I learn from the challenge can be applied in my workplace’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{tw1}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I enjoyed the team social experience’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw2}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I enjoyed the team collaboration experience’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw3}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I could positively contirbute to my team’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw4}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - My efforts were recognised by other team members’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw5}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I found it easy to keep track of my team’s contrbutions’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw6}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I felt that some team members were too dominant’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw7}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I felt that I had to lead to get the job done’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{pf1}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Developing a shared mission amongst team members’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf2}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Working towards a unified goal’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf3}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Managing contributions’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf4}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Keeping track of team decisions’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf5}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Working in a flexible and agile way as analysis progressed’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf6}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Innovative problem solving’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf7}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Enabling an efficient workflow’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf8}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Meeting deadlines’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf9}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Clear communication amongst team members’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf10}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Ability to move easily between engaging with others, or disengaging for independent work’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf11}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Information sharing amongst team members’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf12}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Worked together positively’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf13}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Making decisions’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf14}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Production of useful output’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pfComments}{Response to the prompt: ‘Any additional comments about how the Platform affected teamwork and collaboration?’}
  \item{fb1}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the feedback you received on your team’s submitted reports, do you agree or disagree with the following statements - The feedback accurately reflected the quality of our reports’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{fb2}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the feedback you received on your team’s submitted reports, do you agree or disagree with the following statements - The feedback helped me understand the strengths and weaknesses of our reports’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{fb3}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the feedback you received on your team’s submitted reports, do you agree or disagree with the following statements - The feedback helped me build expertise in analytic reasoning’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{fb4}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the feedback you received on your team’s submitted reports, do you agree or disagree with the following statements - My team used the feedback to improve their submitted reports’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{exCap1}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Analytic report writing’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap2}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Using structured analytic techniques’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap3}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Using OSINT tools and resources’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap4}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Applying strategic thinking frameworks’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap5}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Identifying and analysing assumptions’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap6}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Evaluating quality of analytic reasoning’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap7}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Using decision making frameworks’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{mostValuable}{Response to the prompt: ‘What was the one most valuable thing you learned in the HC2020?’}
  \item{cha1}{Numerical encoding of the participant’s response to the prompt: ‘If you consider participating in an exercise like the HC2020 as analytical skills training, how does it compare with more typical training methods? - A Challenge exercise is more effective’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{cha2}{Numerical encoding of the participant’s response to the prompt: ‘If you consider participating in an exercise like the HC2020 as analytical skills training, how does it compare with more typical training methods? - A Challenge exercise is more engaging’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{cha3}{Numerical encoding of the participant’s response to the prompt: ‘If you consider participating in an exercise like the HC2020 as analytical skills training, how does it compare with more typical training methods? - A Challenge exercise works better for some people, but not as well for others’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{career1}{Numerical encoding of the participant’s response to the prompt: ‘How has participating in the Challenge changed your interest in an intelligence analysis career? My interest has:’. 1 = ‘Not changed’, 2 = ‘Increased’}
  \item{ca1}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about Contending Analyses, do you agree or disagree with the following statements? - I understood what Contending Analyses is’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’}
  \item{ca2}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about Contending Analyses, do you agree or disagree with the following statements? - My team used Contending Analyses’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’}
  \item{ca3}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about Contending Analyses, do you agree or disagree with the following statements? - Using Contending Analyses helped improve the quality of reasoning in reports’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’}
  \item{ca4}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about Contending Analyses, do you agree or disagree with the following statements? - I intend to apply Contending Analyses in my work’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’}
  \item{swarm1}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Platform, do you agree or disagree with the following statements? - Working on the Platform produced better reasoned reports than my normal methods’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Neutral’, 4 = ‘Agree’, 5 = ‘Strongly agree’}
  \item{swarm2}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Platform, do you agree or disagree with the following statements? - My team’s reports were better than I could have produced on my own’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Neutral’, 4 = ‘Agree’, 5 = ‘Strongly agree’}
  \item{swarm3}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Platform, do you agree or disagree with the following statements? - Using a Platform like this would improve intelligence analysis in my organisation’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Neutral’, 4 = ‘Agree’, 5 = ‘Strongly agree’}
  \item{swarm4}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Platform, do you agree or disagree with the following statements? - If my organisation introduced a Platform like this, I would want to use it’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Neutral’, 4 = ‘Agree’, 5 = ‘Strongly agree’}
  \item{lk1}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - It was easy to find the relevant tools’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lk2}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - The tools were well explained’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lk3}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - Overall, the Lens Kit was very helpful’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lk4}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - I preferred the tools already known to me’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lk5}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - The Lens Kit was difficult to navigate’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lk6}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - I personally used it often’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lkSuggestions}{Response to the prompt: ‘Do you have any suggestions to improve the Lens Kit?’}
  \item{res1}{Logical. Indicates whether the participant selected ‘The resources were an important contribution to problem solving’ among their responses to the prompt: ‘Thinking about creating responses (both Resources and Reports) on the platform, Please select all that apply:’.}
  \item{res2}{Logical. Indicates whether the participant selected ‘It was easy to keep track of all the resources posted on the Platform’ among their responses to the prompt: ‘Thinking about creating responses (both Resources and Reports) on the platform, Please select all that apply:’.}
  \item{res3}{Logical. Indicates whether the participant selected ‘My team created a lot of resources’ among their responses to the prompt: ‘Thinking about creating responses (both Resources and Reports) on the platform, Please select all that apply:’.}
  \item{res4}{Logical. Indicates whether the participant selected ‘It was too time consuming to read through everyone's resources’ among their responses to the prompt: ‘Thinking about creating responses (both Resources and Reports) on the platform, Please select all that apply:’.}
  \item{whyRate1}{Logical. Indicates whether the participant selected ‘I used rating to fairly indicate the readiness or quality of a report’ among their responses to the prompt: ‘When you used the Report rating tool, which of the following was your primary purpose?  You can select more than one.’.}
  \item{whyRate2}{Logical. Indicates whether the participant selected ‘I used rating to give guidance to the author’ among their responses to the prompt: ‘When you used the Report rating tool, which of the following was your primary purpose?  You can select more than one.’.}
  \item{whyRate3}{Logical. Indicates whether the participant selected ‘I used rating to push my prefered report to the top’ among their responses to the prompt: ‘When you used the Report rating tool, which of the following was your primary purpose?  You can select more than one.’.}
  \item{ratingTool}{Numerical encoding of the participant’s response to the prompt: ‘Did you use the Report rating tool to rate the Reports your team created?’. 1 = ‘No’, 2 = ‘Yes’}
  \item{ratingToolWhyNot}{Response to the prompt: ‘If you tended not to rate Reports, can you say why?’}
  \item{featureRequests}{Response to the prompt: ‘What additional features or tools would you like to see implemented on the Platform?’}
  \item{externalTools}{Numerical encoding of the participant’s response to the prompt: ‘Did you use any tools to support your work on the Challenge problems that were not included on the Platform?  (eg other software such as Excel, Google Docs, Mapping tools, Chat, Conferencing, etc…)’. 1 = ‘No’, 2 = ‘Yes’}
  \item{externalToolsComments}{Response to the prompt: ‘Please briefly describe these tools and your main reason for using them.’}
  \item{bestQuestionNotAsked}{Response to the prompt: ‘What's the best question that wasn't asked on this survey?’}
  \item{testimonial}{Response to the prompt: ‘Do you have a testimonial about your experience in the HC2020 that you would be happy for us to share publicly? We’ll maintain your anonymity by using your username.’}
  \item{otherComments}{Response to the prompt: ‘Anything else that you would like to say?’}
  \item{aomt}{Actively Open-Minded Thinking score. Formula: aomt1 + aomt2 + aomt3 + (8 - aomt4) + (8 - aomt5) + (8 - aomt6) + (8 - aomt7) + aomt8 + (8 - aomt9) + aomt10 + aomt11}
}
}
\description{
One row per participant who was registered for this platform instance. Most of the columns in this table contain responses to questions on the entry and exit surveys.
}
