% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{parts}
\alias{parts}
\title{parts}
\format{
\describe{
  \item{user}{Username.}
  \item{isOrg}{Logical. Equals TRUE if user participated as part of an organisational team.}
  \item{gaveConsent2}{Gave consent to participate in the ‘Problem Solving In Online Groups’ experiment.}
  \item{gaveConsent3}{Gave consent to participate in the ‘Identifying and Rating Quality of Reasoning’ experiment.}
  \item{agreedToTerms}{}
  \item{agegroup}{Age group.}
  \item{gender}{Gender.}
  \item{occupation}{Occupation status.}
  \item{education}{Highest education level.}
  \item{studyarea}{Study area.}
  \item{studyareaOtherInput}{Free text input for ‘Other’ study area, if selected.}
  \item{int1}{Logical. Indicates whether the participant selected ‘What the problems will be like’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int2}{Logical. Indicates whether the participant selected ‘Platform functionality’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int3}{Logical. Indicates whether the participant selected ‘How reports are created’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int4}{Logical. Indicates whether the participant selected ‘Team collaboration experience’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int5}{Logical. Indicates whether the participant selected ‘The structured training available’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int6}{Logical. Indicates whether the participant selected ‘The tools in the Lens Kit’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int7}{Logical. Indicates whether the participant selected ‘The Contending Analyses methodology’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int8}{Logical. Indicates whether the participant selected ‘The evaluation methods’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int9}{Logical. Indicates whether the participant selected ‘How my team performs’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int10}{Logical. Indicates whether the participant selected ‘Whether the public do as well as the professionals’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{int11}{Logical. Indicates whether the participant selected ‘Other’ among their responses to the prompt: ‘Thinking about the Hunt Challenge, what are you most interested in?’.}
  \item{interestsOtherInput}{If the team selected ‘Other’ among their responses to the prompt ‘Thinking about the Hunt Challenge, what are you most interested in?’, this field contains their free text response.}
  \item{enExpct1}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Interesting problems’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct2}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Reasonable time commitment’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct3}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Difficult problems’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct4}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: I will learn new skills and tools’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct5}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Problems are achievable, for a team’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct6}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: The platform will be a productive work space’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct7}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Training in analytical techniques’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct8}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: Positive team working experience’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct9}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: An effective collaboration, compared to my normal methods’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{enExpct10}{Numerical encoding of the participants' response to the prompt: ‘What are your expectations for your Challenge experience?: What I learn from the challenge can be applied in my workplace’. 1 = ‘I don’t expect this’, 2 = ‘Neutral’, 3 = ‘I do expect this’.}
  \item{pri1}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Seeing how my team performs against the others’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri2}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Testing my own skills and abilities’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri3}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - I think it will be fun’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri4}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Developing my overall analytic skills’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri5}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Learning how the SWARM Platform works’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri6}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Learning about Contending Analyses’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri7}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Trying a new style of team collaboration’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri8}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Contributing to research in intelligence analysis’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri9}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Getting selected for a Super Team’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri10}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Exploring the Lens Kit’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{pri11}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Participating in crowd sourced intelligence analysis’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{priOther}{Numerical encoding of the participants' response to the prompt: ‘How important to you are the following reasons for taking part in the Challenge? - Other’. 1 = ‘Not important’, 2 = ‘Neutral’, 3 = ‘Important’.}
  \item{priOtherInput}{Specified ‘Other’ priority referred to by priOther (if applicable)}
  \item{aomt1}{Numerical encoding of the participants' response to the prompt: ‘Allowing oneself to be convinced by an opposing argument is a sign of good character.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt2}{Numerical encoding of the participants' response to the prompt: ‘People should take into consideration evidence that goes against their beliefs.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt3}{Numerical encoding of the participants' response to the prompt: ‘People should revise their beliefs in light of new information or evidence.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt4}{Numerical encoding of the participants' response to the prompt: ‘Changing your mind is a sign of weakness.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt5}{Numerical encoding of the participants' response to the prompt: ‘Intuition is the best guide in making decisions.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt6}{Numerical encoding of the participants' response to the prompt: ‘It is important to persevere in your beliefs even when evidence is brought to bear against them.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt7}{Numerical encoding of the participants' response to the prompt: ‘One should disregard evidence that conflicts with one's established beliefs.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt8}{Numerical encoding of the participants' response to the prompt: ‘People should search actively for reasons why their beliefs might be wrong.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt9}{Numerical encoding of the participants' response to the prompt: ‘When we are faced with a new question, the first answer that occurs to us is usually best.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt10}{Numerical encoding of the participants' response to the prompt: ‘When faced with a new question, we should consider more than one possible answer before reaching a conclusion.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{aomt11}{Numerical encoding of the participants' response to the prompt: ‘When faced with a new question, we should look for reasons why our first answer might be wrong, before deciding on an answer.’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Somewhat disagree’, 4 = ‘Neither agree nor disagree’, 5 = ‘Somewhat agree’, 6 = ‘Agree’, 7 = ‘Strongly agree’.}
  \item{enCap1}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Analytic report writing’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap2}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Using structured analytic techniques’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap3}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Using OSINT tools and resources’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap4}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Applying strategic thinking frameworks’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap5}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Identifying and analysing assumptions’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap6}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Evaluating quality of analytic reasoning’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{enCap7}{Numerical encoding of the participants' response to the prompt: ‘How would you assess your current capability in the follow areas? This may be from your job, from studies, or from other activities. - Using decision making frameworks’. 1 = ‘None’, 2 = ‘Low’, 3 = ‘Moderate’, 4 = ‘High’.}
  \item{yearsWorkExperience}{Response to the prompt: ‘How many years work experience do you have?’}
  \item{yearsAnalyticalExperience}{Response to the prompt: ‘How many years experience do you have analysing complex problems or data?’}
  \item{ae1}{Logical. Indicates whether the participant selected ‘No direct experience’ among their responses to the prompt: ‘Do you have experience analysing complex problems or data? Please select all applicable.’.}
  \item{ae2}{Logical. Indicates whether the participant selected ‘Yes, in an intelligence or related field’ among their responses to the prompt: ‘Do you have experience analysing complex problems or data? Please select all applicable.’.}
  \item{ae3}{Logical. Indicates whether the participant selected ‘Yes, in a scientific field’ among their responses to the prompt: ‘Do you have experience analysing complex problems or data? Please select all applicable.’.}
  \item{ae4}{Logical. Indicates whether the participant selected ‘Yes, in another field’ among their responses to the prompt: ‘Do you have experience analysing complex problems or data? Please select all applicable.’.}
  \item{ae5}{Logical. Indicates whether the participant selected ‘Prefer not to say’ among their responses to the prompt: ‘Do you have experience analysing complex problems or data? Please select all applicable.’.}
  \item{hasMultidisciplinaryExperience}{Logical. Indicates whether the participant responded ‘Yes’ to the prompt: ‘Have you had experience working in a multidisciplinary team? (i.e. composed of members with varied but complementary experience, qualifications, and skills that contribute to the achievement of a specific objective)’}
  \item{multidisciplinaryExperienceInput}{Response to the prompt: ‘Please describe your experience working in a multidisciplinary team:’}
  \item{starRating}{Integer. Number of stars selected in response to the prompt: ‘Please rate your overall Challenge experience out of 5 stars:’}
  \item{timeWellSpent}{Numerical encoding of the participants' response to the prompt: ‘Did you feel that your participation was time well spent?’. 1 = ‘No’, 2 = ‘Unsure’, 3 = ‘Yes’}
  \item{bestThing}{Response to the prompt: ‘What was the one BEST thing about your Challenge experience?’}
  \item{worstThing}{Response to the prompt: ‘What was the one WORST thing about your Challenge experience?’}
  \item{hoursPerWeek}{Response to the prompt: ‘How many hours did your organisation allow for your participation during work hours per week? Please enter the number, 1, 2, 3, etc...’}
  \item{enoughtTime}{Numerical encoding of the participants' response to the prompt: ‘Did your organisation allow you enough time off regular duties to participate fully?’. 1 = ‘No’, 2 = ‘Yes’}
  \item{proportionOwnTime}{Response to the prompt: ‘What proportion of your participation was done in your own time? (outside of work hours)? Please enter the number, 10, 25, 50, etc...’}
  \item{rate1}{Numerical encoding of the participants' response to the prompt: ‘How do you rate these aspects of the Challenge? - The onboarding process’. 1 = ‘Poor’, 2 = ‘Average’, 3 = ‘Good’}
  \item{rate2}{Numerical encoding of the participants' response to the prompt: ‘How do you rate these aspects of the Challenge? - Our communication with you’. 1 = ‘Poor’, 2 = ‘Average’, 3 = ‘Good’}
  \item{rate3}{Numerical encoding of the participants' response to the prompt: ‘How do you rate these aspects of the Challenge? - The training’. 1 = ‘Poor’, 2 = ‘Average’, 3 = ‘Good’}
  \item{rate4}{Numerical encoding of the participants' response to the prompt: ‘How do you rate these aspects of the Challenge? - The feedback’. 1 = ‘Poor’, 2 = ‘Average’, 3 = ‘Good’}
  \item{rate5}{Numerical encoding of the participants' response to the prompt: ‘How do you rate these aspects of the Challenge? - The Help Center’. 1 = ‘Poor’, 2 = ‘Average’, 3 = ‘Good’}
  \item{exExpct1}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Interesting problems’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct2}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Reasonable time commitment’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct3}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Difficult problems’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct4}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: I will learn new skills and tools’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct5}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Problems are achievable, for a team’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct6}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: The platform will be a productive work space’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct7}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Training in analytical techniques’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct8}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: Positive team working experience’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct9}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: An effective collaboration, compared to my normal methods’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{exExpct10}{Numerical encoding of the participant’s response to the prompt: ‘Given your expectations prior to the Challenge, how did we do?: What I learn from the challenge can be applied in my workplace’. 1 = ‘Below’, 2 = ‘Met’, 3 = ‘Exceeded’, 4 = ‘I had no expectations’.}
  \item{tw1}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I enjoyed the team social experience’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw2}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I enjoyed the team collaboration experience’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw3}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I could positively contirbute to my team’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw4}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - My efforts were recognised by other team members’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw5}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I found it easy to keep track of my team’s contrbutions’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw6}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I felt that some team members were too dominant’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{tw7}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the teamwork aspect of the Challenge, do you agree or disagree with the following statements? - I felt that I had to lead to get the job done’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{pf1}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Developing a shared mission amongst team members’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf2}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Working towards a unified goal’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf3}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Managing contributions’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf4}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Keeping track of team decisions’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf5}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Working in a flexible and agile way as analysis progressed’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf6}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Innovative problem solving’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf7}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Enabling an efficient workflow’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf8}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Meeting deadlines’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf9}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Clear communication amongst team members’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf10}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Ability to move easily between engaging with others, or disengaging for independent work’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf11}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Information sharing amongst team members’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf12}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Worked together positively’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf13}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Making decisions’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pf14}{Numerical encoding of the participant’s response to the prompt: ‘Do you think the Platform supported you in the following areas? - Production of useful output’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’.}
  \item{pfComments}{Response to the prompt: ‘Any additional comments about how the Platform affected teamwork and collaboration?’}
  \item{fb1}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the feedback you received on your team’s submitted reports, do you agree or disagree with the following statements - The feedback accurately reflected the quality of our reports’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{fb2}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the feedback you received on your team’s submitted reports, do you agree or disagree with the following statements - The feedback helped me understand the strengths and weaknesses of our reports’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{fb3}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the feedback you received on your team’s submitted reports, do you agree or disagree with the following statements - The feedback helped me build expertise in analytic reasoning’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{fb4}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the feedback you received on your team’s submitted reports, do you agree or disagree with the following statements - My team used the feedback to improve their submitted reports’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{exCap1}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Analytic report writing’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap2}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Using structured analytic techniques’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap3}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Using OSINT tools and resources’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap4}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Applying strategic thinking frameworks’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap5}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Identifying and analysing assumptions’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap6}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Evaluating quality of analytic reasoning’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{exCap7}{Numerical encoding of the participant’s response to the prompt: ‘After participating in the Challenge, do you think your capability has improved in the following areas? - Using decision making frameworks’. 1 = ‘No’, 2 = ‘Somewhat’, 3 = ‘Significantly’.}
  \item{mostValuable}{Response to the prompt: ‘What was the one most valuable thing you learned in the HC2020?’}
  \item{cha1}{Numerical encoding of the participant’s response to the prompt: ‘If you consider participating in an exercise like the HC2020 as analytical skills training, how does it compare with more typical training methods? - A Challenge exercise is more effective’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{cha2}{Numerical encoding of the participant’s response to the prompt: ‘If you consider participating in an exercise like the HC2020 as analytical skills training, how does it compare with more typical training methods? - A Challenge exercise is more engaging’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{cha3}{Numerical encoding of the participant’s response to the prompt: ‘If you consider participating in an exercise like the HC2020 as analytical skills training, how does it compare with more typical training methods? - A Challenge exercise works better for some people, but not as well for others’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’.}
  \item{career1}{Numerical encoding of the participant’s response to the prompt: ‘How has participating in the Challenge changed your interest in an intelligence analysis career? My interest has:’. 1 = ‘Not changed’, 2 = ‘Increased’}
  \item{ca1}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about Contending Analyses, do you agree or disagree with the following statements? - I understood what Contending Analyses is’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’}
  \item{ca2}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about Contending Analyses, do you agree or disagree with the following statements? - My team used Contending Analyses’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’}
  \item{ca3}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about Contending Analyses, do you agree or disagree with the following statements? - Using Contending Analyses helped improve the quality of reasoning in reports’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’}
  \item{ca4}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about Contending Analyses, do you agree or disagree with the following statements? - I intend to apply Contending Analyses in my work’. 1 = ‘No’, 2 = ‘Yes’, 3 = ‘Not sure’}
  \item{swarm1}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Platform, do you agree or disagree with the following statements? - Working on the Platform produced better reasoned reports than my normal methods’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Neutral’, 4 = ‘Agree’, 5 = ‘Strongly agree’}
  \item{swarm2}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Platform, do you agree or disagree with the following statements? - My team’s reports were better than I could have produced on my own’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Neutral’, 4 = ‘Agree’, 5 = ‘Strongly agree’}
  \item{swarm3}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Platform, do you agree or disagree with the following statements? - Using a Platform like this would improve intelligence analysis in my organisation’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Neutral’, 4 = ‘Agree’, 5 = ‘Strongly agree’}
  \item{swarm4}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Platform, do you agree or disagree with the following statements? - If my organisation introduced a Platform like this, I would want to use it’. 1 = ‘Strongly disagree’, 2 = ‘Disagree’, 3 = ‘Neutral’, 4 = ‘Agree’, 5 = ‘Strongly agree’}
  \item{lk1}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - It was easy to find the relevant tools’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lk2}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - The tools were well explained’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lk3}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - Overall, the Lens Kit was very helpful’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lk4}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - I preferred the tools already known to me’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lk5}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - The Lens Kit was difficult to navigate’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lk6}{Numerical encoding of the participant’s response to the prompt: ‘Thinking about the Lens Kit, do you agree or disagree with the following statements? - I personally used it often’. 1 = ‘Disagree’, 2 = ‘Neutral’, 3 = ‘Agree’}
  \item{lkSuggestions}{Response to the prompt: ‘Do you have any suggestions to improve the Lens Kit?’}
  \item{res1}{Logical. Indicates whether the participant selected ‘The resources were an important contribution to problem solving’ among their responses to the prompt: ‘Thinking about creating responses (both Resources and Reports) on the platform, Please select all that apply:’.}
  \item{res2}{Logical. Indicates whether the participant selected ‘It was easy to keep track of all the resources posted on the Platform’ among their responses to the prompt: ‘Thinking about creating responses (both Resources and Reports) on the platform, Please select all that apply:’.}
  \item{res3}{Logical. Indicates whether the participant selected ‘My team created a lot of resources’ among their responses to the prompt: ‘Thinking about creating responses (both Resources and Reports) on the platform, Please select all that apply:’.}
  \item{res4}{Logical. Indicates whether the participant selected ‘It was too time consuming to read through everyone's resources’ among their responses to the prompt: ‘Thinking about creating responses (both Resources and Reports) on the platform, Please select all that apply:’.}
  \item{whyRate1}{Logical. Indicates whether the participant selected ‘I used rating to fairly indicate the readiness or quality of a report’ among their responses to the prompt: ‘When you used the Report rating tool, which of the following was your primary purpose?  You can select more than one.’.}
  \item{whyRate2}{Logical. Indicates whether the participant selected ‘I used rating to give guidance to the author’ among their responses to the prompt: ‘When you used the Report rating tool, which of the following was your primary purpose?  You can select more than one.’.}
  \item{whyRate3}{Logical. Indicates whether the participant selected ‘I used rating to push my prefered report to the top’ among their responses to the prompt: ‘When you used the Report rating tool, which of the following was your primary purpose?  You can select more than one.’.}
  \item{ratingTool}{Numerical encoding of the participant’s response to the prompt: ‘Did you use the Report rating tool to rate the Reports your team created?’. 1 = ‘No’, 2 = ‘Yes’}
  \item{ratingToolWhyNot}{Response to the prompt: ‘If you tended not to rate Reports, can you say why?’}
  \item{featureRequests}{Response to the prompt: ‘What additional features or tools would you like to see implemented on the Platform?’}
  \item{externalTools}{Numerical encoding of the participant’s response to the prompt: ‘Did you use any tools to support your work on the Challenge problems that were not included on the Platform?  (eg other software such as Excel, Google Docs, Mapping tools, Chat, Conferencing, etc…)’. 1 = ‘No’, 2 = ‘Yes’}
  \item{externalToolsComments}{Response to the prompt: ‘Please briefly describe these tools and your main reason for using them.’}
  \item{bestQuestionNotAsked}{Response to the prompt: ‘What's the best question that wasn't asked on this survey?’}
  \item{testimonial}{Response to the prompt: ‘Do you have a testimonial about your experience in the HC2020 that you would be happy for us to share publicly? We’ll maintain your anonymity by using your username.’}
  \item{otherComments}{Response to the prompt: ‘Anything else that you would like to say?’}
  \item{aomt}{Actively Open-Minded Thinking score. Formula: aomt1 + aomt2 + aomt3 + (8 - aomt4) + (8 - aomt5) + (8 - aomt6) + (8 - aomt7) + aomt8 + (8 - aomt9) + aomt10 + aomt11}
}
}
\description{
One row per participant who was registered for this platform instance. Most of the columns in this table contain responses to questions on the entry and exit surveys.
}
